fix: redirect grep output to /dev/null.
This is needed for recent versions of Apache server.  The output of
grep was sent back as a result of the SOAP request.  The output was
considered as a HTML header.  This header was ignored by previous
versions of Apache, but it contained a whitespace character.  Those
were always forbidden, but this error was ignored by old version of
Apache.